handle exception when two neighbors are simultaneously clicked
This solution keeps the game from entering an unplayable state, but does not prevent this mode for cheating. i.e. simultaneous clicks are not handled with stateful accuracy, meaning it is possible to use simultaneous clicks to skirt correct on/off flipping behavior. This fix also ensures that reset can be reached at all times